Move __main__.report_outcome to app.App.report_outcome
This is part of rendering the app's output, so it makes sense to move it into App.
